requirements.txt
openenv-core>=0.2.3
fastapi>=0.135.0
uvicorn>=0.42.0
pydantic>=2.12.0
huggingface_hub>=1.8.0
openai>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
python-dotenv>=1.0.0